Implemented method where the parameter lists are being processed into files and given out via the frontend as downloads
This allows the parameters to be stored locally and the parameters to be used a little bit more in the backend.

Also implemented the global open subtree button

This button allows to open the entirety of the parameters on the page.

Blocked enter-key-event in the searchbar

This was done so that the page would not reload, while the debounce takes place with the search. This is so that the user does not stop the entire process and remove his / her progress.