Added Utility.js with some useful functions which might be required by other classes or functions in the future
For that to shape up, some more general functions have been removed from the SynchronousScrollerUtility.js and some more have been added too.

ParameterBranchDisplay.js and ParameterDisplay.js received big updates

The synchronous clicking of the nodes has now been activated, making the separate synchronous clicking of the ParameterBranchDisplay.js obsolete, so now there is synchronous clicking through the keys (in the comparison view) and that mechanism also handles the synchronous subtree opening of the ParameterBranchDisplay.js. For that to be possible, new helper functions have been added to the Utility.js class.