Rewrote the ParameterLocationRetrieval in both the frontend and (mostly) the backend and added both that it is now displayed what siteaccess the parameter value comes from and which of the values is actually active
Then the problem with the "\" was fixed

This was done especially because it would hinder every other functionality which existed on the site (as css could not read the backslash at all). Therefore both the keys and values as well as the inline values separately are checked for the backslash and the slash is subsequently being removed from the attribute in order to ensure that all functions can work with the nodes as intended.