Added the ability to include any custom parameter subtree the user wants in the site access view
For that reason there have been changes to the ConfigProcessController.php, which has basically only seen the consequences of the renaming of the ControllerUtility.php to Utility.php (because the class is now also used by the ConfigProcessCoordinator.php).

With the TwigConfigDisplayService.php it is the same and it has also only seen the renaming of a function taking place.

Renamed ControllerUtility.php to Utility.php

This was done because the ConfigProcessCoordinator.php now also employes the Utility and its functions.

Scrapped unnecessary function in the ConfigProcessCoordinator.php

This function was useless, since the exact same thing plus a lot more was being done by another function in the same class, so this was removed.

Also added the custom parameter handling in addition to the existing site access parameters

As requested and stated before, this was done to allow users to add additional subtrees to the site access list and view and it therefore employs a separate class which handles all that in conjunction with the Utility.php.

CustomSiteAccessParamProcessor.php was added

It does what it says and scans the given parameters for such which might be site access dependent and returns the processed parameter trees in order for them to be given to the frontend.

Small Refactorings for the ProcessedParamModel.php and SiteAccessParamProcessor.php to make the code in there more precise and more visible.