Updated Utility.php and removed the CustomParameter-Functions
This was done because there was no need for this functionality to be separate from the main custom-parameter-functionality in the CustomSiteAccessParamProcessor.php.

Added the removed functions from Utility.php into CustomSiteAccessParamProcessor.php

This was done to consolidate all the functionality regarding the custom parameters into one class and bundle them here, as it makes way more sense to collect the functions that way.

ConfigProcessCoordinator.php received a small update as the scanning for potential site access parameters can now be deactivated

Due to the fact that the scouring of the parameters for potential site access dependencies sometimes led to parameter keys being removed even though the parameters were definitely not site access dependent (the default key is a good example of a key which is going to be used outside the site access context and is going to be removed subsequently because of it also appearing in the site accesses), the option to disable that feature has now been added.