PHP:
- ConfigProcessController.php has seen a small addition in which the download denominator in sent in all cases which improves the overall consistency in the methodology

- TwigHelpParserService.php has been added

This class is responsible for parsing certain markdown files which are present in the doc directory of the bundle. Then the parsed (html) output is being given to the template and added there. This is done so that it is possible to display a help text for every part of the bundle in the frontend, which improves the overall usability a lot.

- TwigTestDisplayService.php has seen a few experimental changed, which have been reverted immediately

- ParametersToFileWriter.php has seen a complete refactoring and subsequent overhaul

In order to improve the quality of the code and make the intent of the class and the code more clear, readable and precise, the code formerly present in just one method, has been extracted to multiple dedicated methods only responsible for handling a certain part of the file writing. This should overall improve the readability of the classes functions enormously. However this process is not yet complete as some of the code might be improved in later commits.

- Parsedown.php has been added

The Parsedown.php class has been added to the bundle. This bundle stems from https://parsedown.org/ and can be employed as a standalone bundle. However, we elected to include the file directly in the bundle itself in order to improve the consistency and security of the file (to ensure no future releases of the bundle break the functionality and to have the option to edit the code). Since it is only one file, this is acceptable as introducing an entire dependency just for one class would not be beneficial.

- Utility.php has seen a minor refactoring and a bug fix

The "RemoveCommonParameters" - function had a problem with handling differences when it came to one parameter being of type array and its counterpart not being an array. Since that is a case which already qualifies parameters as different (therefore no further processing has to take place), a new check has been introduced to avoid errors and overall improve the consistency and reliability of the function.

Config:

- routing.yaml: The location retrieval controller has been renamed within the yaml configuration to better fit its new place in the bundle structure, therefore a route has been edited to accommodate for that

- services.yml: Here the controller has first been renamed, at the same time, the new TwigHelpParserService.php has been introduced as a service in the file

Documentation:

- A new help directory has been added in the documentation directory to introduce more precise documentation

In this new directory help files will be placed which are supposed to provide help texts within the actual frontend and more help documentation in general. This is so that a better user experience can be achieved as most pitfalls will be explained and can be avoided in the future.

CSS:

- bundle.css has seen some adaptations for the new help display in every view of the frontend of the bundle, overall these changes have been made so that the help can be accessed easily and a few changes to previously existing elements were made for better access and visibility

- github-markdown-css.css has been added from https://cdnjs.com in order to format the newly added help-markdown display in a good and helpful way and make the overall look of the help more familiar to the user

 JS:

 - DownloadParametersUtility.js has seen a small bug fix:

 Previously when a site access context had been selected for the favourite-view and one went to download the parameters, the site access parameters would be brought to a file instead of the favourites and this has now been fixed.

 - FavourButtonUtility.js: A new visual indication for favourites being removed has been added and unfavoured parameters will are now removed when in the dedicated favourite view to make the view feel more like one would expect and only display the current favourites

 - The ScrollUpButtonDisplay.js has been renamed to its current name due it now also handling the help buttons which are present in every view of the bundle-frontend

 This is due to the fact that the help button and its help-display-functionality has been added to every view of the bundle and since that is the case, the button will need to be setup in every view. Since the Utility didn't have a huge responsibility beforehand, this new one has been given to it.

 - SearchBarUtility.js: The clear-input button is now hidden when the searchbar is empty and will appear, once text is being put into the searchbar

 This was done to avoid having a useless button present in the bar at all times and to make it look better and cleaner when it is empty.

 - FavouritesHandlingUtility.js has seen the introduction of an experimental function which was created to combat an minor issue which is still present after this commit, the function will likely be removed in the next commit

 - app.js has been adapted to the new help-display and sets this up now too

 Views:

 - For most of the parameter views, the displayed eZPlatform header has been outsourced to the base layout and now there are only additions through dedicated template blocks from within the dedicated templates, since the overall header does not really change from view to view.

 - At the same time, the new help text display has been added to every view (even though the help resources are not complete as of now)

 - A few new descriptions in the form of titles have been added to a few elements too, all in an effort to make the offered elements in the bundle more clear