Commit regarding issues #5, #6 and #7
No in depth comments on every class, since no real code changes have occurred.

The biggest changes include removing all scalar type hints from method signatures (since these do not seem to be present in 7.1) and adding documentation to every class.

The documentation includes descriptions of methods, attributes and the class itself.

Also the coding style was brought more inline with Symfony's coding standards.

All of this was done to lower the requirements for the bundle and make the code more understandable and readable.